requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.4
django-appconf==1.0.4
django-compressor==2.4.1
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-debug-toolbar==3.2.1
django-libsass==0.8
django-modelcluster==5.1
django-rest-swagger==2.2.0
django-summernote==0.8.11.6
django-taggit==1.4.0
django-templated-mail==1.1.1
django-treebeard==4.5.1
djangorestframework==3.12.4
djangorestframework-jsonapi==4.2.0
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
draftjs-exporter==4.1.2
gunicorn==20.1.0
html5lib==1.1
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
libsass==0.21.0
Markdown==3.3.4
MarkupSafe==2.0.1
olefile==0.46
openapi-codec==1.3.2
Pillow==8.2.0
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2021.1
rcssmin==1.0.6
requests==2.25.1
rjsmin==1.1.0
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.5
wagtail==2.13.1
wagtailmodeladmin==2.5.7
webencodings==0.5.1
Willow==1.4
